🔧 Wrench: 알림 요청 및 FCM 토큰 발급 시점 변경
